feat: add interceptor for custom response
